% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdpee.R
\name{kdpee}
\alias{kdpee}
\title{D. Stowell and M. D. Plumbley
Fast multidimensional entropy estimation by k-d partitioning.
IEEE Signal Processing Letters 16 (6), 537--540, June 2009.
http://dx.doi.org/10.1109/LSP.2009.2017346}
\usage{
kdpee(X, z = 1.96)
}
\arguments{
\item{X}{data}

\item{z}{Z-score threshold}
}
\value{
Entropy estimate
}
\description{
D. Stowell and M. D. Plumbley
Fast multidimensional entropy estimation by k-d partitioning.
IEEE Signal Processing Letters 16 (6), 537--540, June 2009.
http://dx.doi.org/10.1109/LSP.2009.2017346
}
\examples{
Xu <- matrix(runif(1000 * 100), ncol=100)
kdpee(Xu)

Xn <- matrix(rnorm(1000 * 100), ncol=100)
kdpee(Xn)

}
